+++ lib/webpay_rails.rb
@@ -1,3 +1,4 @@
+require 'rails'
 require 'active_support/concern'
 
 require 'signer'
@@ -13,9 +14,14 @@ require 'webpay_rails/soap'
 require 'webpay_rails/verifier'
 require 'webpay_rails/transaction'
 require 'webpay_rails/transaction_result'
+require 'webpay_rails/railites'
 
 module WebpayRails
   autoload :Base, 'webpay_rails/base'
+
+  class << self
+    attr_accessor :logger
+  end
 
   def self.extended(base)
     base.include WebpayRails::Base