+++ lib/webpay_rails/railites.rb
@@ -1,0 +1,7 @@
+module WebpayRails
+  class Railties < ::Rails::Railtie
+    initializer 'Rails logger' do
+      WebpayRails.logger = Rails.logger
+    end
+  end
+end